Refactored modal files
