+++ client/src/pages/dropdown/friends.tsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react";
 import Navbar from "../navbar";
 import "../../styles/dropdown/friends.css";
 import defaultAvatar from "../../assets/user.png";
-import Modal from "../modal"; 
-import ConfirmModal from "../confirmModal";
+import Modal from "../modal/modal"; 
+import ConfirmModal from "../modal/confirmModal";
 
 const Friends = () => {
   /* =============================== State initializations =================================== */
