+++ client/src/pages/modal/confirmModal.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../styles/modal.css";
+import "../../styles/modal.css";
 
 interface ConfirmModalProps {
   message: string;
